apify-actors/Content_Blueprint/sentiment-compass-ai-powered/requirements.txt
apify-client >= 1.2.0
requests
google-ads
protobuf
python-dateutil
openai
aiohttp